v2/pyproject.toml
[project]
name = "twitch-cli"
version = "0.0.1"
requires-python = ">=3.13,<4"
dependencies = [
    "argcomplete (>=3.6.3,<4.0.0)",
    "xdg-base-dirs (>=6.0.2,<7.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "prettytable (>=3.17.0,<4.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]

[project.scripts]
twitch = "twitch_cli.cli:main"

[tool.poetry.group.test.dependencies]
pytest = "^9.0.1"

[build-system]
requires = ["poetry-core>=2.2.1,<3.0.0"]
build-backend = "poetry.core.masonry.api"
